feat: promote the fuzz oracle bridge to WasmTarget.Bridge
The type-directed, bit-exact JS↔wasm value transport (descriptors,
accessor/ctor closures compiled alongside targets, generic JS
walker/builder, exact-integer comparators) moves from test/fuzz into
the package as WasmTarget.Bridge, so downstream consumers
(PlutoIslands.jl bond marshalling) share the fuzzer-proven
implementation. FuzzBridge/FuzzBridgeArgs become thin wrappers keeping
only the harness runners; their round-trip suites pass unchanged
(26/26, 15/15). New: 'fields' descriptors carry field names so
name-keyed sources (msgpack objects) can map onto ctor args.
